Whocares
My name is Dan and I’ve been an English teacher for many years now. I’ve always been fascinated with learning new things and a while ago I said to myself that I should learn programming before it was too late. I strongly believe that code is something we’ll have to know eventually if we are to remain relative in the workmarket. For example, many of my students are programmers and more often than not I have to either help them with some technical terms, which I need to study in order to know, or look at a coding sheet where I get completely lost. Taking that into consideration, I believe that’s why I’m here and learning how to code.